Bugfix: cycle counter overflow
It seems the cycle counters on the device are 32-bit register which silently
overflow. Added overflow-handling.

+++ src/Simulation/CUDA/kernel/CycleCounters.hpp
@@ -32,7 +32,7 @@ struct CycleCounters
 
 		size_t m_partitionCount;
 
-		int m_clockRateKHz;
+		unsigned long long m_clockRateKHz;
 
 		bool m_stdpEnabled;
